+++ src/core/data.jl
@@ -118,6 +118,13 @@ function update_data(data::Dict{String,Any}, new_data::Dict{String,Any})
 end
 
 
+"calls the replicate function with PowerModels' global keys"
+function replicate(sn_data::Dict{String,Any}, count::Int; global_keys::Set{String}=Set{String}())
+    pm_global_keys = Set(["time_elapsed", "baseMVA", "per_unit"])
+    return InfrastructureModels.replicate(sn_data, count, global_keys=union(global_keys, pm_global_keys))
+end
+
+
 ""
 function apply_func(data::Dict{String,Any}, key::String, func)
     if haskey(data, key)